pyproject.toml
[project]
name = "hw-mqtt-srv"
version = "0.1.0"
description = "MQTT AI Agent server for streaming speech-to-speech communication with IoT devices"
readme = "README.md"
authors = [
    { name = "MQTT AI Agent", email = "agent@example.com" }
]
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.0.0",
    "openai>=1.30.0",
    "aiohttp>=3.9.0",
    "asyncio-mqtt>=0.16.0",
    "websockets>=12.0",
    "pydub>=0.25.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "uvloop>=0.19.0; sys_platform != 'win32'",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.8.0",
]

[project.scripts]
mqtt-ai-server = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
